# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pharma Flow API",
        "description": "API documentation for the Pharma Flow project",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get a list of all orders for the current user or all orders for staff/admin",
                "operationId": "a74d17c8a6f11cd4c219f370004ea189",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderPagination"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Create a new order from the user's cart",
                "operationId": "8da2918aaa60371aa3392e0c612e06bf",
                "requestBody": {
                    "description": "Order data, including optional prescriptions and subscribe type",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "subscribe_type": {
                                        "type": "string",
                                        "enum": [
                                            "none",
                                            "weekly",
                                            "monthly"
                                        ],
                                        "example": "none"
                                    },
                                    "delivery_type": {
                                        "type": "string",
                                        "enum": [
                                            "basic",
                                            "rapid",
                                            "emergency"
                                        ],
                                        "example": "basic"
                                    },
                                    "prescription_images[]": {
                                        "description": "Array of prescription image files",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "description": "Order data without prescription images",
                                "properties": {
                                    "subscribe_type": {
                                        "type": "string",
                                        "enum": [
                                            "none",
                                            "weekly",
                                            "monthly"
                                        ],
                                        "example": "none"
                                    },
                                    "delivery_type": {
                                        "type": "string",
                                        "enum": [
                                            "basic",
                                            "rapid",
                                            "emergency"
                                        ],
                                        "example": "basic"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Order created successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cart not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErrorResponse": {
                "title": "Error Response",
                "description": "Standard error response format for generic errors (e.g., 401, 403, 404, 500)",
                "properties": {
                    "errors": {
                        "description": "Error message",
                        "type": "string"
                    }
                },
                "type": "object",
                "example": {
                    "errors": "Something went wrong."
                }
            },
...
            "Order": {
                "title": "Order",
                "description": "Order model",
                "properties": {
                    "id": {
                        "description": "Order ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "user_id": {
                        "description": "ID of the user who placed the order",
                        "type": "integer",
                        "format": "int64"
                    },
                    "order_date": {
                        "description": "Date and time of the order",
                        "type": "string",
                        "format": "date-time"
                    },
                    "total_amount": {
                        "description": "Total amount of the order",
                        "type": "number",
                        "format": "float"
                    },
                    "order_status": {
                        "description": "Current status of the order",
                        "type": "string",
                        "default": "pending",
                        "enum": [
                            "pending",
                            "delivered",
                            "canceled"
                        ]
                    },
                    "payment_status": {
                        "description": "Current status of the payment",
                        "type": "string",
                        "default": "pending",
                        "enum": [
                            "pending",
                            "paid",
                            "failed"
                        ]
                    },
                    "subscribe_type": {
                        "description": "Current status of the subscription",
                        "type": "string",
                        "default": "none",
                        "enum": [
                            "none",
                            "weekly",
                            "monthly"
                        ]
                    },
                    "created_at": {
                        "description": "Timestamp of order creation",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Timestamp of last update",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object",
                "example": {
                    "id": 1,
                    "user_id": 10,
                    "order_date": "2023-01-01T12:00:00.000000Z",
                    "total_amount": 25.75,
                    "order_status": "pending",
                    "created_at": "2023-01-01T12:00:00.000000Z",
                    "updated_at": "2023-01-01T12:00:00.000000Z"
                }
            },
...
            "OrderPagination": {
                "title": "Order Pagination",
                "description": "Paginated list of orders",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Order"
                        }
                    },
                    "links": {
                        "description": "Pagination links",
                        "type": "object"
                    },
                    "meta": {
                        "description": "Pagination meta information",
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}